requirements.txt
pyrogram~=2.0.104
python-dotenv==1.0.0
TgCrypto==1.2.5
config==0.5.1
Pillow~=9.5.0
moviepy==1.0.3
imageio~=2.28.0
opencv-python~=4.7.0.72
pymongo==4.3.3
decouple~=0.0.7
feedparser~=6.0.10
python-dateutil~=2.8.2
pidfile~=0.1.1
requests~=2.29.0
beautifulsoup4~=4.12.2
html-to-json